Revert "2021_11_24 TimeSchedule 화면 네비게이터 작업"
