+++ Front-end/LimeFront/.gitignore
@@ -2,10 +2,6 @@
 #
 .DS_Store
 
-#files
-package-lock.json
-*.lock
-
 # Xcode
 #
 build/
