--- Front-end/LimeFront/Food.js
@@ -1,10 +1,0 @@
-import * as React from 'react';
-import { Text, StyleSheet, View } from 'react-native';
-
-export default function Food() {
-  return (
-    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-      <Text>Food!</Text>
-    </View>
-  );
-}
