--- Front-end/LimeFront/Profile.js
@@ -1,10 +1,0 @@
-import * as React from 'react';
-import { Text, Stylesheet, View } from 'react-native';
-
-export default function Profile() {
-  return (
-    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-      <Text>Profile!</Text>
-    </View>
-  );
-}
