+++ Front-end/LimeFront/index.js
@@ -5,6 +5,5 @@
 import {AppRegistry} from 'react-native';
 import App from './App';
 import {name as appName} from './app.json';
-import 'react-native-gesture-handler';
 
 AppRegistry.registerComponent(appName, () => App);
