+++ Front-end/LimeFront/package.json
@@ -10,18 +10,12 @@
     "lint": "eslint ."
   },
   "dependencies": {
-    "@react-navigation/bottom-tabs": "^6.0.9",
-    "@react-navigation/material-top-tabs": "^6.0.6",
     "@react-navigation/native": "^6.0.6",
     "@react-navigation/native-stack": "^6.2.5",
-    "react": "^17.0.2",
+    "react": "17.0.2",
     "react-native": "0.66.3",
-    "react-native-gesture-handler": "^1.10.3",
-    "react-native-reanimated": "^2.2.4",
     "react-native-safe-area-context": "^3.3.2",
-    "react-native-screens": "^3.9.0",
-    "react-native-tab-view": "^3.1.1",
-    "react-native-vector-icons": "^9.0.0"
+    "react-native-screens": "^3.9.0"
   },
   "devDependencies": {
     "@babel/core": "^7.16.0",
